// src/modules/profile/profile/profile.interface.ts
import { Account, Post, Subscription } from '@prisma/client'
import { Prisma } from '@prisma/client'
import { MyAccountDTO, MyProfileDTO } from './profile.dto'

export type AccountWithProfile = Prisma.AccountGetPayload<{
    include: { profiles: true }
}>

export interface ProfileServiceInterface {
    myProfile(
        accountId: string,
        profileId?: string
    ): Promise<MyProfileDTO | string>
    myAccount(accountId: string): Promise<MyAccountDTO>
    userProfile(
        userPid: string,
        accountId?: string
    ): Promise<UserProfileResult | string>
}

interface MinData {
    name: string
    avatarImage: string | null
    coverImage: string | null
}

interface FullData extends MinData {
    offcial: boolean
    shortInfo?: string
    extraInfo?: string
    otherLinks: string[]
    subscribers: number
}

interface UserProfileData {
    fullData?: FullData
    subscription?: Subscription[]
    posts?: Post[]
    likes?: number
    dislikes?: number
    ratio?: number
}

export type UserProfileResult = MinData | UserProfileData

export type ProfilePrivacyStatsInterface = Prisma.ProfileGetPayload<{
    include: { privacy: true; stats: true }
}>